python-simple-web-api/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
python-jose==3.1.0